move 404 files out of components
