--- src/components/NotFound/GuardianLogo.js
@@ -1,13 +1,0 @@
-// @flow
-
-import styled from 'preact-emotion';
-
-import TheGuardianLogoSVG from 'images/the-guardian-logo.svg';
-
-const GuardianLogo = styled(TheGuardianLogoSVG)({
-    display: 'block',
-    height: 'auto',
-    width: '100%',
-});
-
-export default GuardianLogo;
