--- src/pages/NotFound.js
@@ -1,7 +1,0 @@
-// @flow
-
-import NotFound from 'components/NotFound';
-
-const NotFoundPage = () => <NotFound />;
-
-export default NotFoundPage;
